refactor: move code to chat input and section
